+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Rational
 
+# 0.7.0 (2022-11-28)
+
+- [feat]: `.distance()` and `.distanceSq()` methods;
+- [feat]: `angle()` method;
+
 ## 0.6.0 (2022-11-27)
 
 - [feat]: `.equals()` method;